bot/requirements.txt
nonebot2
nonebot-adapter-onebot
fastapi
uvicorn
